requirements.txt
PyQt6==6.11.0
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
packaging==23.2
requests==2.31.0
gspread-formatting==1.1.2
python-dotenv==1.0.0
